requirements.txt
Pillow~=9.2.0
opencv-python~=4.6.0.66